+++ src/main/java/org/fundacionjala/coding/marcoslara/AveragesOfNumbers.java
@@ -1,4 +1,4 @@
-package org.fundacionjala.coding.marcoslara.averagesofnumbers;
+package org.fundacionjala.coding.marcoslara;
 
 /**
  * Created by Marcos Lara on 6/6/2017.
