+++ src/test/java/org/fundacionjala/coding/marcoslara/EvaporatorTest.java
@@ -1,4 +1,4 @@
-package org.fundacionjala.coding.marcoslara.evalulationevaporator;
+package org.fundacionjala.coding.marcoslara;
 
 import org.junit.Test;
