+++ src/test/java/org/fundacionjala/coding/marcoslara/HighestAndLowestTest.java
@@ -1,4 +1,4 @@
-package org.fundacionjala.coding.marcoslara.highestandlowest;
+package org.fundacionjala.coding.marcoslara;
 
 import org.junit.Test;
